second part of fixes
