+++ build.gradle
@@ -31,7 +31,6 @@ subprojects {
         junit5PlatformVersion = '1.8.1'
         hamcrestVersion = '2.2'
         seleniumVersion = '4.0.0'
-        seleniumServerVersion = '4.0.0-alpha-2'
         allureVersion = '2.15.0'
         jacksonVersion = '2.13.0'
         mockitoVersion = '4.0.0'
